LPD-44395 Move the captcha client extension to its own workspace
