+++ workspaces/liferay-recaptcha-workspace/client-extensions/liferay-recaptcha-custom-element/package.json
@@ -7,7 +7,7 @@
 	"devDependencies": {
 		"esbuild": "^0.20.2"
 	},
-	"name": "@liferay/liferay-sample-custom-element-captcha",
+	"name": "@liferay/liferay-recaptcha-custom-element",
 	"private": true,
 	"scripts": {
 		"build": "esbuild src/index.js --outdir=build/static --bundle --entry-names=[name].[hash] --format=esm --external:react --external:react-dom --loader:.js=jsx"
