+++ workspaces/liferay-recaptcha-workspace/client-extensions/liferay-recaptcha-custom-element/src/components/Captcha.js
