+++ workspaces/liferay-recaptcha-workspace/client-extensions/liferay-recaptcha-etc-spring-boot/LCP.json
