+++ workspaces/liferay-recaptcha-workspace/client-extensions/liferay-recaptcha-etc-spring-boot/client-extension.yaml
@@ -1,0 +1,13 @@
+liferay-recaptcha-etc-spring-boot:
+    captchaName: reCAPTCHA CX
+    captchaResponseParameterName: g-recaptcha-response
+    customElementExternalReferenceCode: LXC:liferay-recaptcha-custom-element
+    name: Liferay reCAPTCHA Etc Spring Boot
+    oAuth2ApplicationExternalReferenceCode: liferay-recaptcha-etc-spring-boot-oauth-application-user-agent
+    resourcePath: /recaptcha/validation
+    type: captcha
+liferay-recaptcha-etc-spring-boot-oauth-application-user-agent:
+    .serviceAddress: localhost:58081
+    .serviceScheme: http
+    name: Liferay reCAPTCHA Etc Spring Boot OAuth Application User Agent
+    type: oAuthApplicationUserAgent
