+++ workspaces/liferay-recaptcha-workspace/client-extensions/liferay-recaptcha-etc-spring-boot/src/main/java/com/liferay/recaptcha/RecaptchaSpringBootApplication.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.sample;
+package com.liferay.recaptcha;
 
 import com.liferay.client.extension.util.spring.boot3.ClientExtensionUtilSpringBootComponentScan;
 
@@ -18,10 +18,10 @@ import org.springframework.context.annotation.Import;
  */
 @Import(ClientExtensionUtilSpringBootComponentScan.class)
 @SpringBootApplication
-public class SampleSpringBootApplication {
+public class RecaptchaSpringBootApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(SampleSpringBootApplication.class, args);
+		SpringApplication.run(RecaptchaSpringBootApplication.class, args);
 	}
 
 }
