+++ workspaces/liferay-recaptcha-workspace/client-extensions/liferay-recaptcha-etc-spring-boot/src/main/resources/application-default.properties
@@ -2,7 +2,7 @@
 # Application
 #
 
-liferay.sample.etc.captcha.secret=\
+liferay.recaptcha.etc.spring.boot.secret=\
     6LeIxAcTAAAAAGG-vFI1TnRWxMZNFuojJ4WifJWe
 
 #
@@ -10,7 +10,7 @@ liferay.sample.etc.captcha.secret=\
 #
 
 liferay.oauth.application.external.reference.codes=\
-    liferay-sample-etc-captcha-oauth-application-user-agent
+    liferay-recaptcha-etc-spring-boot-oauth-application-user-agent
 
 #
 # Spring Boot
