+++ workspaces/liferay-recaptcha-workspace/client-extensions/liferay-recaptcha-etc-spring-boot/src/main/resources/application.properties
