--- workspaces/liferay-sample-workspace/client-extensions/liferay-sample-custom-element-captcha/client-extension.yaml
@@ -1,13 +1,0 @@
-assemble:
-    -   from: build/static
-        into: static
-liferay-sample-custom-element-captcha:
-    friendlyURLMapping: liferay-sample-custom-element-captcha
-    htmlElementName: liferay-sample-custom-element-captcha
-    instanceable: false
-    name: Liferay Sample Custom Element Captcha
-    portletCategoryName: category.client-extensions
-    type: customElement
-    urls:
-        -   index.*.js
-    useESM: true
