--- workspaces/liferay-sample-workspace/client-extensions/liferay-sample-etc-captcha/client-extension.yaml
@@ -1,13 +1,0 @@
-liferay-sample-etc-captcha:
-    captchaName: reCAPTCHA CX
-    captchaResponseParameterName: g-recaptcha-response
-    customElementExternalReferenceCode: LXC:liferay-sample-custom-element-captcha
-    name: Liferay Sample Etc Captcha
-    oAuth2ApplicationExternalReferenceCode: liferay-sample-etc-captcha-oauth-application-user-agent
-    resourcePath: /captcha/validation
-    type: captcha
-liferay-sample-etc-captcha-oauth-application-user-agent:
-    .serviceAddress: localhost:58081
-    .serviceScheme: http
-    name: Liferay Sample Etc Captcha OAuth Application User Agent
-    type: oAuthApplicationUserAgent
